pyproject.toml
[tool.poetry]
name = "nonebot_plugin_maimai"
version = "0.3.1"
description= "Maimai DX plugin for NoneBot"
authors = ["Agnes_Digital <Z735803792@163.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/Agnes4m/nonebot_plugin_maimai"
repository = "https://github.com/Agnes4m/nonebot_plugin_maimai"
keywords = ["maimai", "nonebot2", "plugin"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",

]
include = [
    "LICENSE","README.md"
]
 
[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.8.3"
nonebot2 = "^2.0.0"
nonebot-adapter-onebot = ">=2.1.3"
pillow = ">=9.3.0"
httpx = ">=0.22.0"
nonebot-plugin-txt2img = "^0.3.0"
bs4 = "^0.0.1"
 
[tool.poetry.dev-dependencies]
 
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"